pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
black = "^21.9b0"
requests = "^2.26.0"
more-itertools = "^8.11.0"
pytest = "^6.2.5"
selenium = "^4.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
